Fix rename copypasta
