+++ src/stactools/worldclim/__init__.py
@@ -1,14 +1,14 @@
 import stactools.core
 
-from stactools.nrcanlandcover.stac import create_item
-from stactools.nrcanlandcover.cog import create_cog
+from stactools.worldclim.stac import create_item
+from stactools.worldclim.cog import create_cog
 
 stactools.core.use_fsspec()
 
 
 def register_plugin(registry):
-    from stactools.nrcanlandcover import commands
-    registry.register_subcommand(commands.create_nrcanlandcover_command)
+    from stactools.worldclim import commands
+    registry.register_subcommand(commands.create_worldclim_command)
 
 __version__ = '0.1.5'
 """Library version"""
